+++ CHANGELOG.md
@@ -3,6 +3,16 @@ CHANGELOG
 
 0.31.0
 ------
+- Added support for an alternative preview window layout that is activated
+  when the size of the preview window is smaller than a certain threshold.
+  ```sh
+  # If the width of the preview window is smaller than 50 columns,
+  # it will be displayed above the search window.
+  fzf --preview 'cat {}' --preview-window 'right,50%,border-left,<50(up,30%,border-bottom)'
+
+  # Or you can just hide it like so
+  fzf --preview 'cat {}' --preview-window '<50(hidden)'
+  ```
 - Use SGR mouse mode to support larger terminals
 - Bug fixes and improvements